async_api/etl/requirements.txt
python-dotenv==0.20.0
psycopg2-binary
elasticsearch==8.3.1
pydantic==1.9.1
psycopg2-binary
python-dotenv==0.20.0